Remove gh-pages as submodule in public
This is to allow the wercker build step to function. Otherwise there are
issues where the build tries to use the previous version of the gh-pages
for the "public" submodule, but since the gh-pages deploy clears
history (not my favorite strategy, but I'm not rewriting the tool) that
previous version cannot be found and the build will fail.